requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.2
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-import-export==2.7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.1.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tablib==3.2.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
